# swagger.yaml
openapi: 3.0.0
info:
  title: Code Cloud Agents API
  description: '**Supervised AI System with Engineering Lead Supervisor and Cloud
    Assistant**


    Ein intelligentes Supervisor-System, das AI-Agenten überwacht, Risiken bewertet
    und bei kritischen Entscheidungen automatisch eingreift.


    ## Features

    - 🎯 STOP-Score System (0-100 Risikobewertung)

    - 🛡️ Enforcement Gate (HARD STOP bei Score ≥ 70)

    - 📊 Task Management mit Audit-Trail

    - 🔍 Vollständige Audit-Log Dokumentation

    - 💬 AI-Chat mit Multi-Provider Support

    - 🎫 Demo Invite System mit Usage-Limits


    ## Authentication

    Currently all endpoints are open. Authentication (JWT) is planned for future releases.


    ## Rate Limiting

    - Demo endpoints: 5 requests per 15 minutes

    - Other endpoints: No limit (production will add limits)


    ## Base URL

    - Development: `http://localhost:3000`

    - Production: `http://178.156.178.70:3000`

    '
  version: 0.1.0
  contact:
    name: Step2Job GmbH
    email: support@step2job.de
  license:
    name: Proprietary
servers:
- url: http://localhost:3000
  description: Development server
- url: http://178.156.178.70:3000
  description: Production server
...
paths:
...
  /api/demo/invites:
    post:
      tags:
      - Demo
      summary: Create demo invite (Admin only)
      description: 'Create a new demo invite code with specified limits.

        **Requires:** Admin authentication (future implementation)

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - createdBy
              properties:
                createdBy:
                  type: string
                  example: admin@example.com
                maxRedemptions:
                  type: integer
                  default: 1
                  example: 10
                expiresInDays:
                  type: integer
                  default: 30
                  example: 7
                taskLimit:
                  type: integer
                  default: 10
                  example: 50
                messageLimit:
                  type: integer
                  default: 100
                  example: 500
                usdCredit:
                  type: number
                  default: 5.0
                  example: 10.0
                eurCredit:
                  type: number
                  default: 5.0
                  example: 10.0
      responses:
        '201':
          description: Invite created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: inv123
                  code:
                    type: string
                    example: DEMO8CHR
                  maxRedemptions:
                    type: integer
                    example: 10
                  expiresAt:
                    type: string
                    format: date-time
                    example: '2026-01-02T12:00:00.000Z'
                  inviteUrl:
                    type: string
                    example: http://localhost:3000/demo/redeem?code=DEMO8CHR
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      operationId: demo_post_api_demo_invites
  /api/demo/redeem:
    post:
      tags:
      - Demo
      summary: Redeem demo invite code
      description: 'Redeem an invite code to create a demo user account.

        **Rate Limited:** 5 attempts per 15 minutes per IP

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - code
              - username
              - password
              properties:
                code:
                  type: string
                  example: DEMO8CHR
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                  example: johndoe
                password:
                  type: string
                  minLength: 8
                  example: SecurePassword123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    example: user123
                  username:
                    type: string
                    example: johndoe
                  limits:
                    type: object
                    properties:
                      taskLimit:
                        type: integer
                        example: 50
                      messageLimit:
                        type: integer
                        example: 500
                      usdCredit:
                        type: number
                        example: 10.0
                      eurCredit:
                        type: number
                        example: 10.0
        '400':
          description: Invalid request or invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Too many redemption attempts. Please try again later.
      operationId: demo_post_api_demo_redeem
  /api/demo/stats:
    get:
      tags:
      - Demo
      summary: Get demo statistics (Admin only)
      description: 'Returns statistics about demo invites and users.

        **Requires:** Admin authentication (future implementation)

        '
      parameters:
      - name: X-User-ID
        in: header
        schema:
          type: string
        description: Admin user ID
        example: admin
      responses:
        '200':
          description: Demo statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalInvites:
                    type: integer
                    example: 10
                  totalUsers:
                    type: integer
                    example: 25
                  activeUsers:
                    type: integer
                    example: 20
                  totalTasksCreated:
                    type: integer
                    example: 150
                  totalMessagesCreated:
                    type: integer
                    example: 500
      operationId: demo_get_api_demo_stats
  /api/demo/users/{userId}:
    get:
      tags:
      - Demo
      summary: Get demo user usage stats
      description: Returns usage statistics for a specific demo user
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
        description: Demo user ID
        example: user123
      responses:
        '200':
          description: User usage statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    example: user123
                  username:
                    type: string
                    example: johndoe
                  limits:
                    type: object
                    properties:
                      taskLimit:
                        type: integer
                        example: 50
                      messageLimit:
                        type: integer
                        example: 500
                      usdCredit:
                        type: number
                        example: 10.0
                      eurCredit:
                        type: number
                        example: 10.0
                  usage:
                    type: object
                    properties:
                      tasksCreated:
                        type: integer
                        example: 12
                      messagesCreated:
                        type: integer
                        example: 45
                      usdSpent:
                        type: number
                        example: 2.5
                      eurSpent:
                        type: number
                        example: 2.3
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      operationId: demo_get_api_demo_users_userid
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        details:
          type: object
          example: {}
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT-based authentication (future implementation).

        Currently all endpoints are open.

        '
security: []